requirements.txt
flask~=3.1.1
langchain~=0.3.26
langchain-community~=0.3.25
langgraph~=0.4.9
mysql-connector-python~=9.3.0
langchain-core~=0.3.68
requests~=2.32.4
langchain-experimental~=0.3.4
Flask-JWT-Extended~=4.7.1
werkzeug~=3.1.3
python-dotenv~=1.1.0